% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_data_helpers.R
\name{combine_project_ways}
\alias{combine_project_ways}
\title{Project and combine all ways into a single multipolygon.}
\usage{
combine_project_ways(osm_path, wkid, osmconf, osm_fields, sa2)
}
\arguments{
\item{osm_path}{Path to Osmium output file}

\item{wkid}{WKID to project OSM data with}

\item{osmconf}{Path to temporary OSM GDAL configuration file.}

\item{osm_fields}{Fields to extract from OSM data.}
}
\description{
Project and combine all ways into a single multipolygon.
}
