% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_data_helpers.R
\name{combine_pts_mpoly_project_centroids}
\alias{combine_pts_mpoly_project_centroids}
\title{Combine points and multipolygons}
\usage{
combine_pts_mpoly_project_centroids(osm_path, wkid, osmconf, osm_fields)
}
\arguments{
\item{osm_path}{Path to Osmium output file}

\item{wkid}{WKID to project OSM data with}

\item{osmconf}{Path to temporary OSM GDAL configuration file.}

\item{osm_fields}{Fields to extract from OSM data.}
}
\description{
Combine the \code{multipolygon} and \code{points} layers of an OSM layer by taking the multipolygon centroids.
}
