% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{dist_raster}
\alias{dist_raster}
\title{Generate a distance raster}
\usage{
dist_raster(input_pts, rast_data)
}
\arguments{
\item{input_pts}{Candidates to be the nearest neighbour}

\item{rast_data:}{A list of three elements. \code{rast}, the empty global raster; \code{cell_nums} numeric\link{n}, cell numbers of relevant cells (containing addresses) and \code{centroids_mat}, their centroids as x-y coords.}
}
\value{
The global raster, with the value of each cell being the distance to the nearest neighbour.
}
\description{
Calculates the distance to the nearest neighbour input point of each relevant cell.
}
