% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_data_helpers.R
\name{gdal_osm_mod_tmp_ini}
\alias{gdal_osm_mod_tmp_ini}
\title{Modify OSM driver}
\usage{
gdal_osm_mod_tmp_ini(conf_path, input_tags = NULL)
}
\arguments{
\item{input_tags}{One of a character vector of tags to be applied to all sections, a named list of character vectors, with each name corresponding to a section, or \code{NULL}, to reset the INI file to default.}
}
\description{
Update GDAL OSM INI file to determine which tags to use as columns
}
