% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_functions.R
\name{generate_rast_data}
\alias{generate_rast_data}
\title{Generate raster data}
\usage{
generate_rast_data(addresses, cell_size, bounds)
}
\arguments{
\item{addresses}{Projected coordinates of all addresses}

\item{cell_size}{Size of cells to use in raster}

\item{bounds}{Boundaries of raster in projected coordinate system; must be a multiple of cell_size}
}
\value{
A list of three elements: \code{rast}, the empty global raster; \code{cell_nums}, cell numbers of relevant cells (containing addresses) and \code{centroids_mat}, their centroids as x-y coords.
}
\description{
Establish which cells in the raster actually contain addresses, and determine their centroids.
}
