% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_data_helpers.R
\name{generate_wsl2_path}
\alias{generate_wsl2_path}
\title{Generate a WSL2 path}
\usage{
generate_wsl2_path(rel_path)
}
\description{
For Windows only, generate a WSL-friendly path (\verb{/mnt/<drive>/...})
}
