% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_data_helpers.R
\name{get_osm_summary}
\alias{get_osm_summary}
\title{Get OSM summary}
\usage{
get_osm_summary(input_path, tmp_dir = "tmp")
}
\arguments{
\item{input_path}{Path to the OSM file}

\item{tmp_dir}{Path to store temporary OSM GDAL driver config.}
}
\description{
Parses each layer of an OSM file and prints the number of rows and tag availability statistics.
}
