% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{get_relevant_cell_area_estimate}
\alias{get_relevant_cell_area_estimate}
\title{Get area estimates for each cell}
\usage{
get_relevant_cell_area_estimate(
  input_poly,
  rast_data,
  cell_size,
  detection_distance
)
}
\description{
Rasterise the feature class, and extract the approximate intersection area in a circle of radius \code{detection_distance} around
each address.
}
