% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{join_raster_to_addresses}
\alias{join_raster_to_addresses}
\title{Extract the raster value at each address.}
\usage{
join_raster_to_addresses(rast, addresses)
}
\description{
Extract the raster value at each address.
}
