% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_data_helpers.R
\name{osmium_extract}
\alias{osmium_extract}
\title{Osmium extraction}
\usage{
osmium_extract(osm_command, input_path, output_path, print_output = F)
}
\arguments{
\item{osm_command}{Inserted after \code{tag-filter} in the Osmium command.}

\item{input_path}{Path to OSM input file}

\item{output_path}{Path to Osmium output file}

\item{print_output}{If TRUE, print the Osmium output to console}
}
\description{
Uses the osmium tool to extract relevant data from an OSM dump. Osmium isn't properly supported on Windows, so this function calls Osmium through WSL 2.
Ensure that you have a WSL distro with osmium installed as your default distro.
}
