% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{process_adjacent_area}
\alias{process_adjacent_area}
\title{Wrapper function for area processing}
\usage{
process_adjacent_area(
  osm_command,
  osm_fields = "name",
  feature_name,
  detection_distance,
  addresses,
  rast_data,
  tmp_dir,
  osm_raw_input,
  wkid,
  osmconf,
  cell_size,
  sa2
)
}
\arguments{
\item{osm_command}{Passed to \code{osmium_extract} to generate points feature class if \code{osm_extract} is TRUE.}

\item{osm_fields}{Fields to read in from Osmium output if \code{osm_extract} is TRUE.}

\item{feature_name}{Name of feature}

\item{detection_distance}{Radius of overlap area to calculate intersection.}

\item{addresses}{sf of addresses to calculate nearest feature class point to.}

\item{rast_data}{Output of \code{generate_rast_data}}

\item{tmp_dir}{Location to store Osmium OSM output files.}

\item{osm_raw_input}{OpenStreetMap file for Osmium to filter}

\item{wkid}{WKID for projection}

\item{osmconf}{Path to temporary GDAL OSM driver conf file.}

\item{cell_size}{Size of raster cells.}

\item{sa2}{SA2 sf file.}
}
\value{
A list of two elements:

\code{combined_ways}: A single sf containing all areas extracted.

\code{address_areas} : A single-column tibble with order of \code{addresses} of approximate areas of overlap with feature class around each address.
}
\description{
Generates a tibble of addresses, each with the area of the intersection of a feature class extracted from OSM and a circle of radius \code{detection_distance} around the address.
}
