% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{process_nearest_pts}
\alias{process_nearest_pts}
\title{Process nearest points}
\usage{
process_nearest_pts(
  osm_command = NULL,
  osm_fields = "name",
  feature_name,
  filter_func = NULL,
  osm_extract = TRUE,
  sf_to_process = NULL,
  osmconf,
  addresses,
  rast_data,
  tmp_dir,
  osm_raw_input,
  wkid
)
}
\arguments{
\item{osm_command}{Passed to \code{osmium_extract} to generate points feature class if \code{osm_extract} is TRUE.}

\item{osm_fields}{Fields to read in from Osmium output if \code{osm_extract} is TRUE.}

\item{feature_name}{Name of feature}

\item{filter_func}{Optional function that takes in an sf of points of interest and returns a named list of filtered sfs to analyse individually.}

\item{osm_extract}{TRUE if Osmium is to be called, FALSE otherwise.}

\item{sf_to_process}{Points to calculate distance to if \code{osm_extract} is FALSE.}

\item{osmconf}{Path to temporary GDAL OSM driver conf file.}

\item{addresses}{sf of addresses to calculate nearest feature class point to.}

\item{rast_data}{Output of \code{generate_rast_data}}

\item{tmp_dir}{Location to store Osmium OSM output files.}

\item{osm_raw_input}{OpenStreetMap file for Osmium to filter}

\item{wkid}{WKID for projection}
}
\value{
A list of two elements:

\code{pts_list}: A single sf if \code{filter_func} is not defined, giving the points from OSM, or multiple sfs from \code{filter_func} if provided.

\code{address_dists} : A tibble with order of \code{addresses}, with each col corresponding to a pts list and each val being the approx dist from that address to the nearest pt.
}
\description{
Wrapper function for point processing.
Generates a tibble of addresses giving distances to sets of points.
}
