% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{round_take_quantiles}
\alias{round_take_quantiles}
\title{Round all data columns, and create quantile columns.}
\usage{
round_take_quantiles(pts, cols_to_excl, join_cols, n = 100)
}
\description{
Note that for points data, lower numbers corresponds to higher quantiles, since being closer is "better", but for area data, higher numbers correspond to higher quantiles, since more nearby area is "better".
}
